+++ R/helpers.R
@@ -70,8 +70,10 @@ cv <- function(x, na.rm = TRUE) {
 # Coefficient of dispersion, or coefficient of quartile variation (CQV).
 # (Bonett et al., 2006: Confidence interval for a coefficient of quartile variation).
 cqv <- function(x, na.rm = TRUE) {
-  fives <- stats::fivenum(x, na.rm = na.rm)
-  (fives[4] - fives[2]) / (fives[4] + fives[2])
+  # using type 6:
+  # m = p. p[k] = k / (n + 1). Thus p[k] = E[F(x[k])]. This is used by Minitab and by SPSS.
+  quartiles <- stats::quantile(x, probs = c(0.25, 0.75), na.rm = na.rm, type = 6)
+  (quartiles[2] - quartiles[1]) / (quartiles[2] + quartiles[1])
 }
 
 # source: scales::number -> scales::percent