+++ NEWS.md
@@ -1,3 +1,10 @@
+# 1.0.0.9000
+
+* `top_freq()` now correctly selects bottoms items using negative a number for *n*
+* `freq.default()` is now exported for use in other packages
+* All numeric calculation in the header of frequency tables now use the same algorithm as used by Minitab and SPSS (see 'Type 6' on `stats::quantile()`)
+
+
 # 1.0.0
 
 First release